add favorites and watchlist funtion

+++ js/app.js
@@ -32,19 +32,39 @@ upcomingBtn.addEventListener("click", function() {
 let moviesContainer = document.getElementById("movies-container")
 
 moviesContainer.addEventListener("click", (e) => {
-    if(e.target.classList.contains("fav-btn")) {
-    let card = e.target.closest(".movie-card")
-    let movieId = card.dataset.id
-    getMovieDetails(movieId).then(data => {
-        let favorites = JSON.parse(localStorage.getItem("favorites")) || []
-        favorites.push(data)
-        localStorage.setItem("favorites", JSON.stringify(favorites))
-        alert("Added to favorites!")
-    })
-} 
-    else{
-        let movieId = e.target.closest(".movie-card").dataset.id
-    window.location.href = `details.html?id=${movieId}`
-    }
     
+    if(e.target.classList.contains("fav-btn")) {
+        let card = e.target.closest(".movie-card")
+        let movieId = card.dataset.id
+        getMovieDetails(movieId).then(data => {
+            let favorites = JSON.parse(localStorage.getItem("favorites")) || []
+            let alreadyAdded = favorites.some(movie => movie.id === data.id)
+            if(alreadyAdded) {
+                alert("Already in favorites!")
+            } else {
+                favorites.push(data)
+                localStorage.setItem("favorites", JSON.stringify(favorites))
+                alert("Added to favorites!")
+            }
+        })
+
+    } else if(e.target.classList.contains("watch-list")) {
+        let card = e.target.closest(".movie-card")
+        let movieId = card.dataset.id
+        getMovieDetails(movieId).then(data => {
+            let watchList = JSON.parse(localStorage.getItem("watchList")) || []
+            let alreadyAdded = watchList.some(movie => movie.id === data.id)
+            if(alreadyAdded) {
+                alert("Already in watchList!")
+            } else {
+                watchList.push(data)
+                localStorage.setItem("watchList", JSON.stringify(watchList))
+                alert("Added to watchList!")
+            }
+        })
+
+    } else {
+        let movieId = e.target.closest(".movie-card").dataset.id
+        window.location.href = `details.html?id=${movieId}`
+    }
 })
